backend/requirements.txt
Django==4.2.24
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.1
python-dotenv==1.1.1
bcrypt==5.0.0
cryptography==46.0.1
requests==2.32.5
Pillow==11.3.0
psycopg2-binary==2.9.10
dj-database-url==2.1.0